pyproject.toml
[tool.poetry]
name = "finance-automator-public"
version = "0.1.0"
description = ""
authors = ["Seo, Jang-won <nanpuhaha@gmail.com>"]
readme = "README.md"
license = "GNU AFFERO GENERAL PUBLIC LICENSE"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
streamlit = "^1.38.0"
pandas = "^2.2.2"
msoffcrypto-tool = "^5.4.2"
openpyxl = "^3.1.5"
xlsxwriter = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
